==> scripts/02_exploration_halfhourly.Rmd <==
---
title: "Explore half-hourly data"
author: "Jessica Guo"
date: "4/2/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Soil data from RainMan SR

After relating daily mean soil water content and soil water potential, we now focus on the S1-S4 treatment period of 7/13/2020 to 9/30/2020 and examine the half-hourly data

```{r}
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(lubridate)
library(cowplot)
```

## Half-hourly data exploration

As before, we load in the half-hourly time series data for VWC, which contains VWC and WP. We match VWC and VWP up for the S1-S4 period, including only plots 3_4 and 3_7. 

```{r}
tm <- data.frame(Period = c("Winter 2020", "Premonsoon", "S1-S4", "Winter 2021"),
                     st = c(as.Date("2019-11-01"), as.Date("2020-05-01"), 
                            as.Date("2020-07-13"), as.Date("2020-10-01")),
                     en = c(as.Date("2020-04-30"), as.Date("2020-07-12"),
                            as.Date("2020-09-30"), as.Date("2021-03-06")))
tm$Period <- factor(tm$Period, levels = c("Winter 2020", "Premonsoon", "S1-S4", "Winter 2021"))

WC <- read.csv("../data/Halfhourly_VWC_Weather_030621.csv") %>%
  dplyr::select(-X, -T_inside, -RH_inside, -T_outside, -RH_outside) %>%
  mutate(TIMESTAMP = as.POSIXct(TIMESTAMP),
         date = as.Date(TIMESTAMP)) %>%
  rename(dt = TIMESTAMP) %>%
  tidyr::pivot_longer(!c(dt, date), 
                      names_to = c("Variable", "Plot", "Depth", "House"), 
                      names_pattern = "(.*)_P(.*)_(.*)_H(.*)") %>%
  mutate(House = factor(as.numeric(House)),
         Depth = factor(as.numeric(Depth)),
         Plot = factor(as.numeric(Plot)),
         ID = paste0(Variable, "_P", Plot, "_", Depth, "_H", House)) %>%
  arrange(ID, dt) %>%
  dplyr::select(-ID) %>%
  filter(House == "3" & Plot %in% c("4", "7")) %>%
  tidyr::drop_na() %>%
  tidyr::pivot_wider(names_from = "Variable", values_from = "value") %>%
  mutate(SWP = SWP/1000,
         Period = case_when(date <= tm$en[1] & date >= tm$st[1] ~ tm$Period[1],
                            date <= tm$en[2] & date >= tm$st[2] ~ tm$Period[2],
                            date <= tm$en[3] & date >= tm$st[3] ~ tm$Period[3],
                            date <= tm$en[4] & date >= tm$st[4] ~ tm$Period[4]),
         depth = Depth,
         Depth = case_when(depth == 1 ~ "0-12 cm",
                           depth == 2 ~ "25 cm",
                           depth == 3 ~ "75 cm"),
         year = year(dt),
         week = isoweek(dt))

# Isoweek 1 often spans 2 years, amend "year" column to the larger of the 2 years
WC$year[which(WC$week == 1 & WC$Period == "Winter 2020")] <- 2020
WC$year[which(WC$week == 1 & WC$Period == "Winter 2021")] <- 2021

WC <- WC %>%
  mutate(year_week = paste0(year, "_", week))
```

For plotting purposes, we also note when irrigation occurred. H3P4 received the S1 treatment, and H3P7 received the S4 treatment. 
```{r}
irig <- readr::read_csv("../data/irrigation.csv") %>%
  mutate(dt = lubridate::mdy_h(paste0(date, " 12"), tz = "America/Phoenix")) %>%
  rename(Plot_4 = S1, Plot_7 = S4) %>%
  tidyr::pivot_longer(!c(dt, date), 
                      names_to = "Plot", 
                      names_pattern = "Plot_(.*)") %>%
  rename(irig_mm = value) %>%
  mutate(irig_tf = ifelse(irig_mm > 0, TRUE, FALSE),
         year = year(dt),
         week = isoweek(dt),
         year_week = paste0(year, "_", week),)

# Replicate 3 times with 3 depths
irig_df <- do.call("rbind", replicate(3, irig, simplify = FALSE))
irig_df$Depth <- c(rep("0-12 cm", nrow(irig)),
                   rep("25 cm", nrow(irig)),
                   rep("75 cm", nrow(irig)))


save(WC, file = "../soildata_app/soilWC-WPHourly.Rdata")
save(irig_df, file = "../soildata_app/irrigDaily.Rdata")

```

Half-hourly data plots
```{r, warning = FALSE}
cols <- brewer.pal(11, "Spectral")
yws <- unique(WC$year_week) # 72
for(yw in yws) {
  # Subset data by week
    sub_wc <- WC %>%
      filter(year_week == yw)
    sub_irig <- irig_df %>%
      filter(year_week == yw & irig_tf == TRUE)
    Year <- unique(sub_wc$year)
    Wk <- sprintf("%02d", unique(sub_wc$week))
    
    # Calculate axis ratios
    prop <- diff(range(sub_wc$SWP))/diff(range(sub_wc$VWC))
    swp_min <- min(sub_wc$SWP)
    
    # Double axis plot
    if(nrow(sub_irig) == 0) {
      fig1 <- ggplot() +
      geom_point(data = sub_wc, aes(x = dt, y = SWP, color = "SWP (MPa)"), size = 0.5) +
      geom_point(data = sub_wc, aes(x = dt,y = VWC*prop + swp_min, color = "VWC"), size = 0.5) +
      facet_grid(rows = vars(Depth), cols = vars(Plot)) +
      scale_x_datetime(date_labels = "%b %d") +
      scale_y_continuous(sec.axis = 
                           sec_axis(~./prop - swp_min/prop,
                                    name = expression(paste(Theta, " (", m^3, " ", m^-3, ")")))) +
      scale_color_manual(values = cols[c(2, 10)]) +
      labs(y = expression(paste(Psi, " (MPa)")), x = "", title = Year) +
      theme_bw(base_size = 12) +
      theme(axis.title.y.left = element_text(color = cols[2]),
            axis.title.y.right = element_text(color = cols[10])) +
      guides(color = FALSE)
    } else {
      fig1 <- ggplot() +
      geom_point(data = sub_irig, aes(x = dt, y = 0), pch = 8) +
      geom_point(data = sub_wc, aes(x = dt, y = SWP, color = "SWP (MPa)"), size = 0.5) +
      geom_point(data = sub_wc, aes(x = dt,y = VWC*prop + swp_min, color = "VWC"), size = 0.5) +
      facet_grid(rows = vars(Depth), cols = vars(Plot)) +
      scale_x_datetime(date_labels = "%b %d") +
      scale_y_continuous(sec.axis = 
                           sec_axis(~./prop - swp_min/prop,
                                    name = expression(paste(Theta, " (", m^3, " ", m^-3, ")")))) +
      scale_color_manual(values = cols[c(2, 10)]) +
      labs(y = expression(paste(Psi, " (MPa)")), x = "", title = Year) +
      theme_bw(base_size = 12) +
      theme(axis.title.y.left = element_text(color = cols[2]),
            axis.title.y.right = element_text(color = cols[10])) +
      guides(color = FALSE)
    }
    
    
    fig2 <- ggplot(sub_wc) +
      geom_point(aes(x = VWC, y = SWP, color = Depth), size = 0.5) +
      facet_wrap(~Plot, ncol = 2) +
      scale_color_manual(values = cols[c(3, 7, 9)]) +
      labs(y = expression(paste(Psi, " (MPa)")), 
           x = expression(paste(Theta, " (", m^3, " ", m^-3, ")"))) +
      theme_bw(base_size = 12) +
      guides(color = FALSE)

    ggsave2(paste0("../plots/weekly/", Year, "_", Wk, ".jpg"),
            plot = plot_grid(fig1, fig2, nrow = 2, rel_heights = c(2, 1)),                             
            height = 6, width = 6, units = "in")

    
    print(paste0("Done with ", yw))
}
```
